+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 #### Added
 - Social Login support for Google and Facebook
 - Biometric Authentication with WebAuthn
+- Exposed Revoke access token method [SDKS-980] - 'FRUser.getCurrentUser().revokeAccessToken(Listener)'
+
 
 # Version 2.2.0